[RED] Postcode location service should save a postcode with coordinates

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/domain/Postcode.kt
@@ -16,10 +16,10 @@ data class Postcode(
   var code: String,
 
   @Column(name = "x_coordinate", nullable = false)
-  var xCoordinate: Float,
+  var xCoordinate: Double,
 
   @Column(name = "y_coordinate", nullable = false)
-  var yCoordinate: Float,
+  var yCoordinate: Double,
 ) : Auditable() {
   override fun toString(): String = """
     Postcode(
